Fix travis errors
